Bringing fork up to date

+++ setup.py
@@ -1,41 +1,47 @@
-from distutils.core import setup
+#!/usr/bin/env python2
+# -*- coding=utf-8 -*-
+
+from __future__ import print_function
+
 import sys
 import os
 
-import S3.PkgInfo
-
-if float("%d.%d" % sys.version_info[:2]) < 2.4:
-    sys.stderr.write("Your Python version %d.%d.%d is not supported.\n" % sys.version_info[:3])
-    sys.stderr.write("S3cmd requires Python 2.4 or newer.\n")
-    sys.exit(1)
-
 try:
-    import xml.etree.ElementTree as ET
-    print "Using xml.etree.ElementTree for XML processing"
-except ImportError, e:
+    import xml.etree.ElementTree
+    print("Using xml.etree.ElementTree for XML processing")
+except ImportError as e:
     sys.stderr.write(str(e) + "\n")
     try:
-        import elementtree.ElementTree as ET
-        print "Using elementtree.ElementTree for XML processing"
-    except ImportError, e:
+        import elementtree.ElementTree
+        print("Using elementtree.ElementTree for XML processing")
+    except ImportError as e:
         sys.stderr.write(str(e) + "\n")
         sys.stderr.write("Please install ElementTree module from\n")
         sys.stderr.write("http://effbot.org/zone/element-index.htm\n")
         sys.exit(1)
 
-try:
-    ## Remove 'MANIFEST' file to force
-    ## distutils to recreate it.
-    ## Only in "sdist" stage. Otherwise
-    ## it makes life difficult to packagers.
-    if sys.argv[1] == "sdist":
+from setuptools import setup
+
+import S3.PkgInfo
+
+if float("%d.%d" % sys.version_info[:2]) < 2.6:
+    sys.stderr.write("Your Python version %d.%d.%d is not supported.\n" % sys.version_info[:3])
+    sys.stderr.write("S3cmd requires Python 2.6 or newer.\n")
+    sys.exit(1)
+
+## Remove 'MANIFEST' file to force
+## distutils to recreate it.
+## Only in "sdist" stage. Otherwise
+## it makes life difficult to packagers.
+if len(sys.argv) > 1 and sys.argv[1] == "sdist":
+    try:
         os.unlink("MANIFEST")
-except:
-    pass
+    except OSError as e:
+        pass
 
 ## Re-create the manpage
 ## (Beware! Perl script on the loose!!)
-if sys.argv[1] == "sdist":
+if len(sys.argv) > 1 and sys.argv[1] == "sdist":
     if os.stat_result(os.stat("s3cmd.1")).st_mtime < os.stat_result(os.stat("s3cmd")).st_mtime:
         sys.stderr.write("Re-create man page first!\n")
         sys.stderr.write("Run: ./s3cmd --help | ./format-manpage.pl > s3cmd.1\n")
@@ -47,7 +53,7 @@ if not os.getenv("S3CMD_PACKAGING"):
     man_path = os.getenv("S3CMD_INSTPATH_MAN") or "share/man"
     doc_path = os.getenv("S3CMD_INSTPATH_DOC") or "share/doc/packages"
     data_files = [
-        (doc_path+"/s3cmd", [ "README", "INSTALL", "NEWS" ]),
+        (doc_path+"/s3cmd", [ "README.md", "INSTALL", "NEWS" ]),
         (man_path+"/man1", [ "s3cmd.1" ] ),
     ]
 else:
@@ -65,6 +71,8 @@ setup(
     ## Packaging details
     author = "Michal Ludvig",
     author_email = "michal@logix.cz",
+    maintainer = "github.com/mdomsch, github.com/matteobar, github.com/fviard",
+    maintainer_email = "s3tools-bugs@lists.sourceforge.net",
     url = S3.PkgInfo.url,
     license = S3.PkgInfo.license,
     description = S3.PkgInfo.short_description,
@@ -74,7 +82,29 @@ setup(
 Authors:
 --------
     Michal Ludvig  <michal@logix.cz>
-""" % (S3.PkgInfo.long_description)
-    )
+""" % (S3.PkgInfo.long_description),
+
+    classifiers = [
+        'Development Status :: 5 - Production/Stable',
+        'Environment :: Console',
+        'Environment :: MacOS X',
+        'Environment :: Win32 (MS Windows)',
+        'Intended Audience :: End Users/Desktop',
+        'Intended Audience :: System Administrators',
+        'License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)',
+        'Natural Language :: English',
+        'Operating System :: MacOS :: MacOS X',
+        'Operating System :: Microsoft :: Windows',
+        'Operating System :: POSIX',
+        'Operating System :: Unix',
+        'Programming Language :: Python :: 2.6',
+        'Programming Language :: Python :: 2.7',
+        'Programming Language :: Python :: 2 :: Only',
+        'Topic :: System :: Archiving',
+        'Topic :: Utilities',
+    ],
+
+    install_requires = ["python-dateutil", "python-magic"]
+)
 
 # vim:et:ts=4:sts=4:ai
